Initial commit with in-progress pin scheduler project
pinscheduler is a PlatformIO project intended to be deployed on ESP8266
microcontrollers which allows scheduling activation of GPIO pins,
including with a duty cycle (e.g. cycles of 6s on / 4s off until total on
time is 10s).

Current state: Supports hardcoded activation schedules or reading a
set of schedules from /pinschedule.json on disk.